PreProd commit

+++ src/main/java/com/tsimura/repository/PhotoRepository.java
@@ -13,7 +13,7 @@ public interface PhotoRepository extends PagingAndSortingRepository<Photo, Integ
 
     Integer countByUserId(int userId);
 
-    Integer countByOwnerId(int ownerId);
+    Long countByOwnerId(int ownerId);
 
     @Query("select count(distinct p.ownerId) from Photo p where p.userId = :userId")
     Integer groupsCountByUserId(@Param("userId") int userId);
